pyproject.toml
[project]
name = "multiagent"
version = "0.1.0"
description = "mutliagent system"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.2",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-groq>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
